Fixed renaming for collapsed procedure and variable blocks.
