+++ appinventor/blocklyeditor/src/blocks/components.js
@@ -13,7 +13,7 @@
 
 goog.provide('Blockly.Blocks.components');
 goog.provide('Blockly.ComponentBlock');
-
+goog.require('Blockly.Blocks.Utilities');
 goog.require('Blockly.TypeBlock');
 
 /*
@@ -170,7 +170,7 @@ Blockly.Blocks.component_event = {
     if (this.instanceName == oldname) {
       this.instanceName = newname;
       this.componentDropDown.setValue(this.instanceName);
-      Blockly.ComponentBlock.renameCollapsed(this);
+      Blockly.Blocks.Utilities.renameCollapsed(this, 0);
     }
   },
   renameVar: function(oldName, newName) {
@@ -341,7 +341,7 @@ Blockly.Blocks.component_method = {
       //var title = this.inputList[0].titleRow[0];
       //title.setText('call ' + this.instanceName + '.' + this.methodType.name);
       this.componentDropDown.setValue(this.instanceName);
-      Blockly.ComponentBlock.renameCollapsed(this);
+      Blockly.Blocks.Utilities.renameCollapsed(this, 0);
     }
   },
   getMethodTypeObject : function() {
@@ -561,7 +561,7 @@ Blockly.Blocks.component_set_get = {
       //var title = this.inputList[0].titleRow[0];
       //title.setText(this.instanceName + '.');
       this.componentDropDown.setValue(this.instanceName);
-      Blockly.ComponentBlock.renameCollapsed(this);
+      Blockly.Blocks.Utilities.renameCollapsed(this, 0);
     }
   },
   typeblock : function(){
@@ -673,7 +673,7 @@ Blockly.Blocks.component_component_block = {
       //var title = this.inputList[0].titleRow[0];
       //title.setText(this.instanceName);
       this.componentDropDown.setValue(this.instanceName);
-      Blockly.ComponentBlock.renameCollapsed(this);
+      Blockly.Blocks.Utilities.renameCollapsed(this, 0);
     }
   },
 
@@ -708,15 +708,6 @@ Blockly.ComponentBlock.createComponentDropDown = function(block){
     }
   };
   return componentDropDown;
-}
-
-Blockly.ComponentBlock.renameCollapsed = function(block){
-  if (block.isCollapsed()) {
-    var COLLAPSED_INPUT_NAME = '_TEMP_COLLAPSED_INPUT';
-    block.removeInput(COLLAPSED_INPUT_NAME);
-    var text = block.toString(Blockly.COLLAPSE_CHARS);
-    block.appendDummyInput(COLLAPSED_INPUT_NAME).appendField(text);
-  }
 }
 
 Blockly.ComponentBlock.HELPURLS = {
